fix: eliminate all Clang build warnings in project source
- AppStateBuilder.cpp: move internal snapshot functions into anonymous namespace
- LocaleManager.h: add explicit case Language::en in both switches
- PluginOperationController.h: mark destructor with override
- PerformanceFile.cpp: remove unused varToMetadata
- RecordingSessionController.cpp: rename lambda init-captures to avoid shadowing
- MidiTrace.cpp: fix sign-conversion in array subscript

+++ source/Diagnostics/MidiTrace.cpp
@@ -8,7 +8,7 @@ constexpr auto noteNames
     = std::array<const char*, 12> { "C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B" };
 
 const char* noteNameFromSemitone(int note) {
-    return noteNames[note % 12];
+    return noteNames[static_cast<size_t>(note % 12)];
 }
 
 int octaveFromMidiNote(int note) {
